removed wrong TitleSection
