+++ src/views/PreLaunchAirdrop/components/index.ts
@@ -1,4 +1,4 @@
-export { default as TitleSection } from "./SplashFlow/TitleSection";
+export { default as TitleSection } from "./TitleSection";
 export { default as EligibilityPill } from "./EligibilityPill";
 export { default as CardIcon } from "./CardIcon";
 export { default as CardContent } from "./CardContent";
